Move game_items::Storage class definition to GameItems/Storage.h
